+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "comm",
+  "name": "mc",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
