+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "Comm - ProCoSys",
-  "name": "Comm - ProCoSys",
+  "short_name": "MC - ProCoSys",
+  "name": "MC - ProCoSys",
   "icons": [
     {
       "src": "favicon.ico",
